Finish (and improve) chromecast support
Currently, we're restricting playback on only one device at a time;
if chromecast is connected, then the radio will play on chromecast,
and if it's not, then the playback will happen locally.

I guess this might not be what some people want; for example,  if
they want to play radio both on the chromecast and locally at the
same time for some reason, so we can change that at a later date
after getting some feedback.

There are also some additional improvements that could be made, like
controling the volume on the chromecast player through the website,
but leaving those for some other time.